# firebase/functions/src/api/openapi.yaml
openapi: 3.1.0
info:
  title: Flack API
  version: '1'
  summary: Read your channels, messages and people, and post messages, from scripts, CI and bots.
  license:
    name: MIT
    url: https://github.com/mvergarair/flack/blob/main/LICENSE
  description: |
    The Flack API runs inside your own Flack deployment. Requests go straight to your Firebase
    project; nothing passes through a third party.

    ## Authentication

    Every request needs a **personal API token** in the `Authorization` header:

    ```
    Authorization: Bearer flk_…
    ```

    Create one in Flack: your avatar (bottom left) → **API tokens** → name it, pick the access,
    **Create token**. It's shown once, so keep it in a password manager or your CI's secret store.

    A token **acts as you**: it sees exactly the channels and DMs you're in, and messages it posts
    appear as you, with the same notifications, search and link previews as the app. For a bot,
    invite a dedicated account (e.g. `deploy-bot@yourcompany.com`) and create the token there.

    | Access | Scopes | Allows |
    |---|---|---|
    | Read only | `read` | All `GET` endpoints |
    | Read and post messages | `read`, `write` | Also posting messages and opening DMs |

    Flack stores only a SHA-256 hash of each token. Revoke tokens anytime from **API tokens**;
    they also stop working when their owner is deactivated.

    ## Conventions

    - JSON in and out. Timestamps are **milliseconds since the Unix epoch**, or `null`.
    - Mentions in text use Flack's stored format: `<@USER_ID>` for a person, `<!here>` and
      `<!channel>` for everyone online / everyone in the channel. Markdown works as in the app.
    - Lists are paged with `limit` (1–100, default 50) and a cursor (`nextBefore` / `nextAfter`),
      which is `null` on the last page.
    - About **120 requests per minute** per token; over that you get `429` with `Retry-After`.
    - Channels you can't see return `404`, the same as channels that don't exist.

    ## Errors

    ```json
    { "error": { "code": "not_found", "message": "Channel not found (or you are not a member)." } }
    ```

    `code` is one of `invalid_argument` (400), `unauthenticated` (401), `forbidden` and
    `insufficient_scope` (403), `not_found` (404), `method_not_allowed` (405),
    `failed_precondition` (409), `rate_limited` (429) and `internal` (500).

servers:
  - url: /api/v1
    description: This Flack

security:
  - token: []

tags:
  - name: Me
  - name: Users
  - name: Channels
  - name: Messages
  - name: Direct messages
  - name: Search

paths:
  /me:
    get:
      tags: [Me]
      operationId: getMe
      summary: Who am I
      description: The token's owner and the token's scopes. A good first request to check a token.
      responses:
        '200':
          description: The token's owner.
          content:
            application/json:
              schema:
                type: object
                required: [user, scopes]
                properties:
                  user: { $ref: '#/components/schemas/User' }
                  scopes:
                    type: array
                    items: { type: string, enum: [read, write] }
                    example: [read, write]
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }

  /users:
    get:
      tags: [Users]
      operationId: listUsers
      summary: List people
      description: Everyone active in the workspace, sorted by name.
      responses:
        '200':
          description: Active people.
          content:
            application/json:
              schema:
                type: object
                required: [users]
                properties:
                  users:
                    type: array
                    items: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Unauthenticated' }

  /users/{userId}:
    get:
      tags: [Users]
      operationId: getUser
      summary: Get a person
      description: One person, including deactivated people.
      parameters:
        - { $ref: '#/components/parameters/userId' }
      responses:
        '200':
          description: The person.
          content:
            application/json:
              schema:
                type: object
                required: [user]
                properties:
                  user: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '404': { $ref: '#/components/responses/NotFound' }

  /channels:
    get:
      tags: [Channels]
      operationId: listChannels
      summary: List my channels
      description: Channels and DMs you're a member of, most recently active first.
      responses:
        '200':
          description: Your channels and DMs.
          content:
            application/json:
              schema:
                type: object
                required: [channels]
                properties:
                  channels:
                    type: array
                    items: { $ref: '#/components/schemas/Channel' }
        '401': { $ref: '#/components/responses/Unauthenticated' }

  /channels/{channelId}:
    get:
      tags: [Channels]
      operationId: getChannel
      summary: Get a channel
      description: A channel you're a member of, or any public channel.
      parameters:
        - { $ref: '#/components/parameters/channelId' }
      responses:
        '200':
          description: The channel.
          content:
            application/json:
              schema:
                type: object
                required: [channel]
                properties:
                  channel: { $ref: '#/components/schemas/Channel' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '404': { $ref: '#/components/responses/NotFound' }

  /channels/{channelId}/messages:
    get:
      tags: [Messages]
      operationId: listMessages
      summary: List messages
      description: |
        The channel's messages, **newest first**: top-level messages plus thread replies that were
        also sent to the channel. You must be a member. Page with `before` = the previous `nextBefore`.
      parameters:
        - { $ref: '#/components/parameters/channelId' }
        - { $ref: '#/components/parameters/limit' }
        - name: before
          in: query
          description: Only messages created before this timestamp (ms). Pass the previous page's `nextBefore`.
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: A page of messages.
          content:
            application/json:
              schema:
                type: object
                required: [messages, nextBefore]
                properties:
                  messages:
                    type: array
                    items: { $ref: '#/components/schemas/Message' }
                  nextBefore:
                    type: [integer, 'null']
                    format: int64
                    description: Cursor for the next (older) page; `null` on the last page.
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      tags: [Messages]
      operationId: postMessage
      summary: Post a message
      description: |
        Posts as the token's owner (needs the `write` scope and membership, including for DMs).
        Mentioned people and thread participants are notified as usual, and the message is
        searchable within seconds.

        Send an `Idempotency-Key` to make retries safe: repeating a request with the same key
        returns the first message (`200`) instead of posting again.
      parameters:
        - { $ref: '#/components/parameters/channelId' }
        - name: Idempotency-Key
          in: header
          required: false
          description: Any string up to 200 characters, e.g. `deploy-$GITHUB_RUN_ID`.
          schema: { type: string, maxLength: 200 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text:
                  type: string
                  maxLength: 40000
                  description: Markdown, with `<@USER_ID>`, `<!here>` and `<!channel>` mentions.
                  example: '✅ Deployed **web** to production'
                threadId:
                  type: string
                  description: Reply in this thread (the id of the thread's first message).
                alsoToChannel:
                  type: boolean
                  description: With `threadId`, also show the reply in the channel.
            examples:
              message:
                summary: A message
                value: { text: '✅ Deployed **web** to production' }
              reply:
                summary: A thread reply, also sent to the channel
                value: { text: 'Fixed in 1.2.1', threadId: 'MESSAGE_ID', alsoToChannel: true }
      responses:
        '201':
          description: Posted.
          content:
            application/json:
              schema:
                type: object
                required: [message]
                properties:
                  message: { $ref: '#/components/schemas/Message' }
        '200':
          description: Already posted with this `Idempotency-Key`; returns the original message.
          content:
            application/json:
              schema:
                type: object
                required: [message]
                properties:
                  message: { $ref: '#/components/schemas/Message' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: The channel is archived or the thread doesn't exist.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '429': { $ref: '#/components/responses/RateLimited' }

  /channels/{channelId}/messages/{messageId}:
    get:
      tags: [Messages]
      operationId: getMessage
      summary: Get a message
      parameters:
        - { $ref: '#/components/parameters/channelId' }
        - { $ref: '#/components/parameters/messageId' }
      responses:
        '200':
          description: The message.
          content:
            application/json:
              schema:
                type: object
                required: [message]
                properties:
                  message: { $ref: '#/components/schemas/Message' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '404': { $ref: '#/components/responses/NotFound' }

  /channels/{channelId}/messages/{messageId}/replies:
    get:
      tags: [Messages]
      operationId: listReplies
      summary: List thread replies
      description: A thread's replies, **oldest first**. Page with `after` = the previous `nextAfter`.
      parameters:
        - { $ref: '#/components/parameters/channelId' }
        - { $ref: '#/components/parameters/messageId' }
        - { $ref: '#/components/parameters/limit' }
        - name: after
          in: query
          description: Only replies created after this timestamp (ms). Pass the previous page's `nextAfter`.
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: A page of replies.
          content:
            application/json:
              schema:
                type: object
                required: [replies, nextAfter]
                properties:
                  replies:
                    type: array
                    items: { $ref: '#/components/schemas/Message' }
                  nextAfter:
                    type: [integer, 'null']
                    format: int64
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '404': { $ref: '#/components/responses/NotFound' }

  /dms:
    post:
      tags: [Direct messages]
      operationId: openDm
      summary: Open a DM
      description: |
        Opens the DM with one or more people (up to 8 others), creating it if needed. Needs the
        `write` scope. Then post to it with **Post a message** using the returned channel id.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userIds]
              properties:
                userIds:
                  type: array
                  minItems: 1
                  maxItems: 8
                  items: { type: string }
                  description: The other people in the conversation (active members).
            example: { userIds: ['USER_ID'] }
      responses:
        '201':
          description: Created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DmResult' }
        '200':
          description: Already existed.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DmResult' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }

  /search:
    get:
      tags: [Search]
      operationId: search
      summary: Search messages
      description: |
        The app's search: whole words and word beginnings, accent-insensitive, every word must
        match. Only channels you're a member of. Newest first, 20 per page.
      parameters:
        - name: q
          in: query
          required: true
          description: The words to find.
          schema: { type: string }
          example: release notes
        - name: channelId
          in: query
          description: Only this channel.
          schema: { type: string }
        - name: authorId
          in: query
          description: Only messages by this person.
          schema: { type: string }
        - name: hasFile
          in: query
          description: '`true`: only messages with attachments.'
          schema: { type: boolean }
        - name: after
          in: query
          description: Only messages after this timestamp (ms).
          schema: { type: integer, format: int64 }
        - name: before
          in: query
          description: Pagination cursor; pass the previous page's `nextBefore`.
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: Matching messages.
          content:
            application/json:
              schema:
                type: object
                required: [results, nextBefore]
                properties:
                  results:
                    type: array
                    items: { $ref: '#/components/schemas/SearchResult' }
                  nextBefore: { type: [integer, 'null'], format: int64 }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }

components:
  securitySchemes:
    token:
      type: http
      scheme: bearer
      bearerFormat: flk_…
      description: A personal API token from Flack → your avatar → **API tokens**.

  parameters:
    userId:
      name: userId
      in: path
      required: true
      schema: { type: string }
    channelId:
      name: channelId
      in: path
      required: true
      description: From **List my channels**, or the address bar in Flack (`…/c/CHANNEL_ID`).
      schema: { type: string }
    messageId:
      name: messageId
      in: path
      required: true
      schema: { type: string }
    limit:
      name: limit
      in: query
      description: Page size, 1–100.
      schema: { type: integer, minimum: 1, maximum: 100, default: 50 }

  responses:
    BadRequest:
      description: Bad or missing input.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: { code: invalid_argument, message: '`text` is required.' } }
    Unauthenticated:
      description: No token, a malformed one, or a revoked one.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: { code: unauthenticated, message: 'This token was revoked or never existed.' } }
    Forbidden:
      description: The owner isn't active, or the token lacks the `write` scope.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: { code: insufficient_scope, message: 'This token needs the "write" scope for this endpoint.' } }
    NotFound:
      description: Doesn't exist, or you can't see it.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: { code: not_found, message: 'Channel not found (or you are not a member).' } }
    RateLimited:
      description: Too many requests; wait `Retry-After` seconds.
      headers:
        Retry-After:
          schema: { type: integer }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }

  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum: [invalid_argument, unauthenticated, forbidden, insufficient_scope, not_found, method_not_allowed, failed_precondition, rate_limited, internal]
            message: { type: string }

    User:
      type: object
      required: [id, name, title, email, role, status, photoUrl, timeZone, customStatus]
      properties:
        id: { type: string, description: 'Use in mentions as `<@id>`.', example: uMia }
        name: { type: string, example: Mia Member }
        title: { type: string, example: Backend engineer }
        email: { type: string, format: email, example: mia@example.com }
        role: { type: string, enum: [admin, member] }
        status: { type: string, enum: [active, deactivated] }
        photoUrl: { type: [string, 'null'], format: uri }
        timeZone: { type: [string, 'null'], description: IANA time zone., example: America/Santiago }
        customStatus:
          description: Only while it hasn't expired.
          oneOf:
            - type: 'null'
            - type: object
              required: [emoji, text, expiresAt]
              properties:
                emoji: { type: string, example: 🌴 }
                text: { type: string, example: On vacation }
                expiresAt: { type: [integer, 'null'], format: int64 }

    Channel:
      type: object
      required: [id, type, name, topic, memberIds, archived, createdAt, lastMessageAt]
      properties:
        id: { type: string, example: cEngineering }
        type: { type: string, enum: [public, private, dm] }
        name: { type: string, description: 'Without `#`; empty for DMs (use `memberIds`).', example: engineering }
        topic: { type: string, example: Backend, infra and deploys }
        memberIds: { type: array, items: { type: string } }
        archived: { type: boolean, description: Archived channels are read-only. }
        createdAt: { type: [integer, 'null'], format: int64 }
        lastMessageAt: { type: [integer, 'null'], format: int64 }

    Message:
      type: object
      required: [id, channelId, authorId, text, createdAt, editedAt, threadId, alsoToChannel, replyCount, lastReplyAt, mentions, reactions, attachments, deleted]
      properties:
        id: { type: string }
        channelId: { type: string }
        authorId: { type: string }
        text: { type: string, description: 'Markdown with `<@id>` mentions; empty if deleted.', example: 'Build passed <@uAda>' }
        createdAt: { type: [integer, 'null'], format: int64, example: 1759123456789 }
        editedAt: { type: [integer, 'null'], format: int64 }
        threadId: { type: [string, 'null'], description: "For replies: the thread's first message." }
        alsoToChannel: { type: boolean }
        replyCount: { type: integer, description: For thread starters. }
        lastReplyAt: { type: [integer, 'null'], format: int64 }
        mentions:
          type: array
          items: { type: string }
          description: User ids, plus `!here` / `!channel`.
        reactions:
          type: object
          additionalProperties: { type: array, items: { type: string } }
          description: '`{ userId: [emoji, …] }`'
          example: { uAda: ['👍'] }
        attachments:
          type: array
          description: Metadata only; download files in the app.
          items:
            type: object
            required: [name, size, contentType]
            properties:
              name: { type: string, example: report.pdf }
              size: { type: integer }
              contentType: { type: string, example: application/pdf }
        deleted: { type: boolean, description: Deleted messages with replies stay as placeholders. }

    DmResult:
      type: object
      required: [channel, created]
      properties:
        channel: { $ref: '#/components/schemas/Channel' }
        created: { type: boolean }

    SearchResult:
      type: object
      required: [messageId, channelId, threadId, authorId, createdAt, snippet, hasFile]
      properties:
        messageId: { type: string }
        channelId: { type: string }
        threadId: { type: [string, 'null'] }
        authorId: { type: string }
        createdAt: { type: integer, format: int64 }
        snippet: { type: string, example: 'Finish the upload flow and lock down storage…' }
        hasFile: { type: boolean }